requirements.txt
# Core dependencies
streamlit>=1.20.0
jinja2>=3.0.0
pyjwt>=2.6.0
pyperclip>=1.8.2
pyyaml>=6.0
requests>=2.28.0

# UI enhancements
streamlit-ace>=0.1.1

# Dev dependencies
pytest>=7.0.0
pytest-cov>=4.0.0